Added test of analytical equation for area of overlap
Eric, with help from Xinwei, proposed the analytical soluion digrammed in the
figure "png/venn_overlap_area.png".

I corrected the last term in the equation, wrote the script
"numericalVsAnalytical" to compare the output of this equation to Monte Carlo
results, and saved the resulting plot as "numericalVsAnalytical.pdf".

The last term should be dh = d*sqrt(ra^2 - (d^2 + ra^2 - rb^2 )^2/(4*d^2)));
I fixed this in the PNG file by image manipulation, so it is (more or less)
in Eric's own handwriting. Man, I should work for the CIA.